feat!: improve integrator stability and reliability (#416)
This commit improves parameter choices for the integrators overall. In particular, we now make more similar choices to scipy.integrate.solve_ivp, which should lead to fewer surprises.

BREAKING: It was found out that the currently calculated analytical Jacobian leads to problems when implicit integrators are employed. This affects LSODA, Radau and BDF. **The usage of analytical Jacobians is then disabled for this version until we find a fix for the bug**. Even though it is a breaking chance per se, it is mild, as most of the change happened in places the user had previously no access.